backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI/ML
openai==1.10.0
google-generativeai==0.3.2
sentence-transformers==2.3.1
spacy==3.7.2
transformers==4.37.2
torch==2.1.2
numpy==1.26.3
scikit-learn==1.4.0

# Vector Database
qdrant-client==1.7.3

# External APIs
googlemaps==4.10.0
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.2.0
python-dateutil==2.8.2

# File Handling
python-magic==0.4.27
pillow==10.2.0
PyPDF2==3.0.1

# Task Queue
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Utilities
python-dotenv==1.0.1
pydantic-extra-types==2.4.1
email-validator==2.1.0
phonenumbers==8.13.29

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0
